requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3

# Authentication
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9
django-environ==0.11.2

# GIS
geos==0.2.3

# Real-time (WebSocket only)
channels==4.0.0
channels-redis==4.1.0
uvicorn[standard]==0.24.0
websockets==12.0
redis==5.0.1

# File handling
Pillow==10.1.0
openpyxl==3.1.2

# API Documentation
drf-yasg==1.21.7

# Utilities
python-dateutil==2.8.2